Update to CRUD

+++ src/main/java/com/ibm/winnerapp/squad/repository/SquadRepository.java
@@ -4,7 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ibm.winnerapp.squad.model.Squad;
 
-public interface SquadRepository extends JpaRepository<Squad, Integer> {
+public interface SquadRepository extends JpaRepository<Squad, Long> {
+
  
 }
