+++ src/main/java/com/ibm/winnerapp/ss/repository/SsRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ibm.winnerapp.ss.model.Ss;
 
-public interface SsRepository extends JpaRepository<Ss, Integer> {
+public interface SsRepository extends JpaRepository<Ss, Long> {
  
 }
